moving CountDetails based on PR feedback
