--- common.go
@@ -1,35 +1,0 @@
-package servicebus
-
-import "encoding/xml"
-
-//	MIT License
-//
-//	Copyright (c) Microsoft Corporation. All rights reserved.
-//
-//	Permission is hereby granted, free of charge, to any person obtaining a copy
-//	of this software and associated documentation files (the "Software"), to deal
-//	in the Software without restriction, including without limitation the rights
-//	to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-//	copies of the Software, and to permit persons to whom the Software is
-//	furnished to do so, subject to the following conditions:
-//
-//	The above copyright notice and this permission notice shall be included in all
-//	copies or substantial portions of the Software.
-//
-//	THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-//	IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-//	FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-//	AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-//	LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-//	OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-//	SOFTWARE
-
-// CountDetails has current active (and other) messages for queue/topic.
-type CountDetails struct {
-	XMLName                        xml.Name `xml:"CountDetails"`
-	ActiveMessageCount             *int32   `xml:"ActiveMessageCount,omitempty"`
-	DeadLetterMessageCount         *int32   `xml:"DeadLetterMessageCount,omitempty"`
-	ScheduledMessageCount          *int32   `xml:"ScheduledMessageCount,omitempty"`
-	TransferDeadLetterMessageCount *int32   `xml:"TransferDeadLetterMessageCount,omitempty"`
-	TransferMessageCount           *int32   `xml:"TransferMessageCount,omitempty"`
-}
